more fixes for BigSky Plugin, mostly exports
