+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@automattic/big-sky-agents",
-  "version": "1.1.18",
+  "version": "1.1.19",
   "description": "The Big Sky Agents SDK",
   "repository": {
     "type": "git",
