+++ src/index.js
@@ -33,6 +33,10 @@ export { AgentsProvider, useAgents } from './components/agents-provider';
 export { ChatProvider } from './components/chat-provider';
 export { ToolkitsProvider } from './components/toolkits-provider';
 export { default as ChatHistory } from './components/chat-history.jsx';
+export { default as AskUserComponent } from './components/ask-user.jsx';
+export { default as ConfirmComponent } from './components/confirm.jsx';
+export { default as MessageContent } from './components/message-content.jsx';
+export { default as UserMessageInput } from './components/user-message-input.jsx';
 
 /**
  * Hooks
